// ui/src/generated/types/conflictsResponse.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */
import type { Conflict } from './conflict'

export interface ConflictsResponse {
  /** List of conflicts with the current group operation:
   * adding a device into the group
   * updating the schedule of the group
   */
  conflicts: Conflict[]
  description: string
  success: boolean
}