// ui/src/generated/types/groupPayload.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */
import type { GroupPayloadClass } from './groupPayloadClass'
import type { GroupPayloadState } from './groupPayloadState'

/**
 * Payload object for creating/updating a group
 */
export interface GroupPayload {
  /** Group class; privileged value => debug, bookable, standard */
  class?: GroupPayloadClass
  /**
   * Group Name; default value => generated at runtime
   * @pattern ^[0-9a-zA-Z-_./: ]{1,50}$
   */
  name?: string
  /**
   * Group repetitions; default value => 0
   * @minimum 0
   */
  repetitions?: number
  /** Group starting time (in UTC, conforming to RFC 3339 section 5.6); default value => group creation time */
  startTime?: string
  /** Group state; default value => pending or ready for bookable/standard classes */
  state?: GroupPayloadState
  /** Group expiration time (in UTC, conforming to RFC 3339 section 5.6); default value => startTime + 1 hour */
  stopTime?: string
}